Implement Array class
Array is simply a wrapper around ArrayBuffer. It exists to allow
the creation of various search and sort algorithms. It can also be
used as the backing data structure for other types of data
structure in this repository e.g. Set.

I had started by making Array take elements of a generic type but
I had problems I did not know how to overcome when it came to doing
greater than comparisons of the generic type. For simplicity Array
is typed to contain only Int instances for now.

The initial search and sort implementations follow an imperative
approach. Later on as my knowledge of Scala improves I make take
the opportunity to refactor them to use a functional approach that
is idiomatic to Scala.